requirements.txt
# Requirements for the Shelf-life Predictor app
# Comprehensive shelf-life prediction tool using kinetic modeling
streamlit>=1.20.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
plotly>=5.0.0
requests>=2.25.0
requests_cache>=0.9.0
openmeteo_requests>=0.1.0
retry_requests>=0.1.0
streamlit-authenticator>=0.3.0
pyyaml>=6.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0

